Add files via upload
Initial commit.

Most of the functionality is implemented. The only missing features at this point are a few of the 3D content stores. DAZ Product organization is currently functional in a separate script, and will be integrated soon. Until then, I am making that separate script available in this repository. GoGet is not currently multithreaded, but that is planned for a future update.