Add files via upload
Initial commit.

This script should be run in the same directory as downloaded DAZ 3D content files. It scans the directory for files matching DAZ's naming conventions, creates a directory named in the format "SKU_Product Name", moves the ZIP file(s) into the directory, and also downloads any available images. It first searches the store for product images, then the wiki. If both fail, then a text file is created indicating the unavailability of product images.

This script's functionality will be implemented in the main GoGet class in a future update.